Updated parameters to double datatype

+++ Assets/Script/FireController.cs
@@ -17,14 +17,14 @@ public class FireController : MonoBehaviour
         CalculateDistanceFromInlet();
     }
 
-    float getDistanceFromInlet()
+    double getDistanceFromInlet()
     {
         Parameters.Params parameters = param.parameters;
 
-        return 0.153f - (0.322f * parameters.angle) + (0.396f * parameters.areaOther) + (0.424f * parameters.areaCenter) + (0.0226f * parameters.areaConcentric)
-                + (0.175f * parameters.angle * parameters.angle) + (0.0185f * parameters.areaOther * parameters.angle) - (0.0701f * parameters.areaOther * parameters.areaOther) - (0.251f * parameters.areaCenter * parameters.angle)
-                + (0.179f * parameters.areaCenter * parameters.areaOther) + (0.0150f * parameters.areaCenter * parameters.areaCenter) + (0.0134f * parameters.areaConcentric * parameters.angle) + (0.0296f * parameters.areaConcentric * parameters.areaOther)
-                + (0.0752f * parameters.areaConcentric * parameters.areaCenter) + (0.0192f * parameters.areaConcentric * parameters.areaConcentric);
+        return 0.153 - (0.322 * parameters.angle) + (0.396 * parameters.areaOther) + (0.424 * parameters.areaCenter) + (0.0226 * parameters.areaConcentric)
+                + (0.175 * parameters.angle * parameters.angle) + (0.0185 * parameters.areaOther * parameters.angle) - (0.0701 * parameters.areaOther * parameters.areaOther) - (0.251 * parameters.areaCenter * parameters.angle)
+                + (0.179 * parameters.areaCenter * parameters.areaOther) + (0.0150 * parameters.areaCenter * parameters.areaCenter) + (0.0134 * parameters.areaConcentric * parameters.angle) + (0.0296 * parameters.areaConcentric * parameters.areaOther)
+                + (0.0752 * parameters.areaConcentric * parameters.areaCenter) + (0.0192 * parameters.areaConcentric * parameters.areaConcentric);
     }
 
     public void CalculateDistanceFromInlet()
@@ -37,9 +37,9 @@ public class FireController : MonoBehaviour
         transform.parent.position = new Vector3(rmg.getIntersectionDistance(), transform.parent.position.y, transform.parent.position.z);
     }
 
-    public void SetFireLength(float Xcc)
+    public void SetFireLength(double Xcc)
     {
-        transform.parent.localScale = new Vector3(Mathf.Lerp(0.5f * 1f, 1f, Xcc), Mathf.Lerp(0.25f * maxFireLength, maxFireLength, Xcc), 1);
+        transform.parent.localScale = new Vector3(Mathf.Lerp(0.5f * 1f, 1f, (float)Xcc), Mathf.Lerp(0.25f * maxFireLength, maxFireLength, (float)Xcc), 1);
         xcc_text.text = Xcc.ToString();
     }
 }
